backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
twilio==8.10.2
stripe==7.11.0
razorpay==1.3.0
gunicorn==21.2.0
